allow nodes to the Raft cluster join as non-voter
