+++ physical/raft/raft.go
@@ -1,3 +1,4 @@
+// Copyright (c) 2024 OpenBao a Series of LF Projects, LLC
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
@@ -52,6 +53,9 @@ const (
 	// EnvVaultRaftNonVoter is used to override the non_voter config option, telling Vault to join as a non-voter (i.e. read replica).
 	EnvVaultRaftNonVoter  = "BAO_RAFT_RETRY_JOIN_AS_NON_VOTER"
 	raftNonVoterConfigKey = "retry_join_as_non_voter"
+
+	// NodeNonVoter is a custom node type for non-voters
+	NodeNonVoter = autopilot.NodeType("non-voter")
 )
 
 var getMmapFlags = func(string) int { return 0 }
@@ -155,6 +159,10 @@ type RaftBackend struct {
 	// autopilot features. This will be instantiated in both leader and followers.
 	// However, only active node will have a "running" autopilot.
 	autopilot *autopilot.Autopilot
+
+	// delegate is the autopilot delegate that will be used to interact with the
+	// autopilot library.
+	delegate *Delegate
 
 	// autopilotConfig represents the configuration required to instantiate autopilot.
 	autopilotConfig *AutopilotConfig
@@ -990,14 +998,6 @@ func (b *RaftBackend) SetupCluster(ctx context.Context, opts SetupOpts) error {
 			return fmt.Errorf("raft recovery failed to parse peers.json: %w", err)
 		}
 
-		// Non-voting servers are only allowed in enterprise. If Suffrage is disabled,
-		// error out to indicate that it isn't allowed.
-		for idx := range recoveryConfig.Servers {
-			if recoveryConfig.Servers[idx].Suffrage == raft.Nonvoter {
-				return fmt.Errorf("raft recovery failed to parse configuration for node %q: setting `non_voter` not supported in OpenBao", recoveryConfig.Servers[idx].ID)
-			}
-		}
-
 		b.logger.Info("raft recovery found new config", "config", recoveryConfig)
 
 		err = raft.RecoverCluster(raftConfig, b.fsm, b.logStore, b.stableStore, b.snapStore, b.raftTransport, recoveryConfig)
@@ -1309,7 +1309,7 @@ func (b *RaftBackend) GetConfiguration(ctx context.Context) (*RaftConfigurationR
 }
 
 // AddPeer adds a new server to the raft cluster
-func (b *RaftBackend) AddPeer(ctx context.Context, peerID, clusterAddr string) error {
+func (b *RaftBackend) AddPeer(ctx context.Context, peerID, clusterAddr string, voter bool) error {
 	if err := ctx.Err(); err != nil {
 		return err
 	}
@@ -1322,7 +1322,13 @@ func (b *RaftBackend) AddPeer(ctx context.Context, peerID, clusterAddr string) e
 			return errors.New("raft storage is not initialized")
 		}
 		b.logger.Trace("adding server to raft", "id", peerID)
-		future := b.raft.AddVoter(raft.ServerID(peerID), raft.ServerAddress(clusterAddr), 0, 0)
+		var future raft.IndexFuture
+		if voter {
+			future = b.raft.AddVoter(raft.ServerID(peerID), raft.ServerAddress(clusterAddr), 0, 0)
+		} else {
+			future = b.raft.AddNonvoter(raft.ServerID(peerID), raft.ServerAddress(clusterAddr), 0, 0)
+		}
+
 		return future.Error()
 	}
 
@@ -1330,13 +1336,27 @@ func (b *RaftBackend) AddPeer(ctx context.Context, peerID, clusterAddr string) e
 		return errors.New("raft storage autopilot is not initialized")
 	}
 
-	b.logger.Trace("adding server to raft via autopilot", "id", peerID)
+	var nodeType autopilot.NodeType
+	if voter {
+		nodeType = autopilot.NodeVoter
+	} else {
+		nodeType = NodeNonVoter
+	}
+
+	b.logger.Debug("adding server to raft via autopilot", "id", peerID)
+	if !voter {
+		b.logger.Debug("adding non-voter to raft via autopilot", "id", peerID)
+		err := b.delegate.AddNonVoter(raft.ServerID(peerID))
+		if err != nil {
+			return err
+		}
+	}
 	return b.autopilot.AddServer(&autopilot.Server{
 		ID:          raft.ServerID(peerID),
 		Name:        peerID,
 		Address:     raft.ServerAddress(clusterAddr),
 		RaftVersion: raft.ProtocolVersionMax,
-		NodeType:    autopilot.NodeVoter,
+		NodeType:    nodeType,
 	})
 }
