+++ website/content/api-docs/system/storage/raft.mdx
@@ -52,6 +52,19 @@ leader node.
 
 - `auto_join_port` `(int: 8200)` - Port to be used for `auto_join`.
 
+- `non_voter` `(bool: false)` - If set, will make the server not
+  participate in the Raft quorum, and have it only receive the data replication
+  stream. This can be used to add read scalability to a cluster in cases where a
+  high volume of reads to servers are needed. The default is false.
+
+:::warning
+
+Adding a large number of non-voters to a cluster will impact application
+latency. Make sure to tune the [Raft settings](/docs/configuration/storage/raft/#raft-parameters)
+for your intended use case.
+
+:::
+
 ### Sample payload
 
 ```json
