+++ website/content/docs/commands/operator/raft.mdx
@@ -74,6 +74,21 @@ The following flags are available for the `operator raft join` command.
 
 - `-leader-client-key` `(string: "")` - Client key to authenticate to Raft leader.
 
+- `-non-voter` `(bool: false)` - This flag is used to make the
+  server not participate in the Raft quorum, and have it only receive the data
+  replication stream. This can be used to add read scalability to a cluster in
+  cases where a high volume of reads to servers are needed. The default is false.
+  See [`retry_join_as_non_voter`](/docs/configuration/storage/raft#retry_join_as_non_voter)
+  for the equivalent config option when using `retry_join` stanzas instead.
+
+:::warning
+
+Adding a large number of non-voters to a cluster will impact application
+latency. Make sure to tune the [Raft settings](/docs/configuration/storage/raft/#raft-parameters)
+for your intended use case.
+
+:::
+
 - `-retry` `(bool: false)` - Continuously retry joining the Raft cluster upon
   failures. The default is false.
